requirements.txt
altgraph==0.17.4
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
